Changed file names to reflect universal app terminology used, added doughnut chart for survey results
