--- views/partials/quiz-score.handlebars
@@ -1,6 +1,0 @@
-    <!-- <div style="width: 500px;"><canvas id="dimensions"></canvas></div><br/> -->
-    <div style="width: 800px;"><canvas id="quiz-score"></canvas></div>
-
-    <!-- <script type="module" src="dimensions.js"></script> -->
-    <script type="module" src="../charts/quiz-score.js"></script>
-
