+++ views/partials/stress-journey.handlebars
@@ -1,0 +1,8 @@
+<div>
+    <!-- <div style="width: 500px;"><canvas id="dimensions"></canvas></div><br/> -->
+    <div>Stress Journey Chart should appear here</div>
+    <div style="width: 800px;"><canvas id="stress-journey"></canvas></div>
+    
+    <!-- <script type="module" src="dimensions.js"></script> -->
+    <script type="module" src="../charts/stress-journey.js"></script>
+</div>
