--- views/partials/stress-story.handlebars
@@ -1,7 +1,0 @@
-    <div>
-        <!-- <div style="width: 500px;"><canvas id="dimensions"></canvas></div><br/> -->
-        <div style="width: 800px;"><canvas id="stress-journey"></canvas></div>
-    
-        <!-- <script type="module" src="dimensions.js"></script> -->
-        <script type="module" src="../charts/stress-journey.js"></script>
-    </div>
