+++ views/partials/survey-score.handlebars
@@ -1,0 +1,12 @@
+<div>
+    <!-- <div style="width: 500px;"><canvas id="dimensions"></canvas></div><br/> -->
+    <div>Survey Score Chart should appear here</div>
+    <div style="width: 800px;"><canvas id="survey-score"></canvas></div>
+
+    <div>Survey Doughnut should appear here</div>
+        <div style="width: 800px;"><canvas id="survey-doughnut"></canvas></div>
+
+
+    <!-- <script type="module" src="dimensions.js"></script> -->
+    <script type="module" src="../charts/survey-score.js"></script>
+</div>
